FinalFinal

+++ Assignments/Final Project/Pages/routes.php
@@ -2,18 +2,31 @@
 
 $path = "index.php?page=welcome";
 
-
-$nav=<<<HTML
+if($_SESSION['status']=="ad"){
+    $nav=<<<HTML
     <nav>
         <ul>
-            <a href="index.php?page=addContact">Add Contact&emsp;</a>
-            <a href="index.php?page=deleteContacts">Delete contact(s)&emsp;</a>
             <a href="index.php?page=addAdmin">Add Admin&emsp;</a>
             <a href="index.php?page=deleteAdmins">Delete Admin(s)&emsp;</a>
+            <a href="index.php?page=addContact">Add Contact&emsp;</a>
+            <a href="index.php?page=deleteContacts">Delete contact(s)&emsp;</a>
             <a href="index.php?page=logout">Logout</a>
         </ul>
     </nav>
 HTML;
+}
+else{
+    $nav=<<<HTML
+    <nav>
+        <ul>
+            <a href="index.php?page=addContact">Add Contact&emsp;</a>
+            <a href="index.php?page=deleteContacts">Delete contact(s)&emsp;</a>
+            <a href="index.php?page=logout">Logout</a>
+        </ul>
+    </nav>
+HTML;
+}
+
 
 if(isset($_GET)){
     if($_GET['page'] === "addContact"){
@@ -43,6 +56,11 @@ if(isset($_GET)){
 
     else if($_GET['page'] === "deleteAdmins"){
         require_once('Pages/deleteAdmins.php');
+        $result = init();
+    }
+
+    else if($_GET['page'] === "logout"){
+        require_once('logout.php');
         $result = init();
     }
